refactor: update import paths for calendar utilities and add new helper functions
