+++ frontend/app/dashboard/calendar/hooks/useEventState.ts
@@ -2,7 +2,7 @@
 import { useState } from 'react';
 import { CalendarEvent } from '@/utils/calendar/calendar-types';
 import { EventStateActions } from '../types/eventHooks';
-import { sortEventsByStartTime } from '../utils/eventDataProcessing';
+import { sortEventsByStartTime } from '../../../../utils/calendar/eventDataProcessing';
 
 /**
  * Hook for managing event state
