+++ frontend/utils/calendar/eventDataProcessing.ts
