+++ frontend/utils/calendar/eventValidation.ts
